requirements.txt
django==5.2.8
gunicorn==23.0.0

cloudinary==1.44.1
django-cloudinary-storage==0.3.0

python-dotenv==1.2.1

requests==2.32.5
urllib3==2.5.0
idna==3.11
charset-normalizer==3.4.4
six==1.17.0

tzdata==2025.2
sqlparse==0.5.3
asgiref==3.11.0
typing_extensions==4.15.0
certifi==2025.11.12
